file-based: rename stream file to correct pattern
