+++ airbyte_cdk/sources/file_based/stream/identities_stream.py
@@ -10,7 +10,7 @@ from airbyte_cdk.sources.file_based.discovery_policy import AbstractDiscoveryPol
 from airbyte_cdk.sources.file_based.exceptions import FileBasedErrorsCollector
 from airbyte_cdk.sources.file_based.file_based_stream_reader import AbstractFileBasedStreamReader
 from airbyte_cdk.sources.streams.core import JsonSchema
-from airbyte_cdk.sources.streams.permissions.identitiesstream import IdentitiesStream
+from airbyte_cdk.sources.streams.permissions.identities_stream import IdentitiesStream
 
 
 class FileIdentitiesStream(IdentitiesStream):
